Add testing crate
This moves the testaso macro to the testing crate.
