+++ sgx-types/src/attr.rs
@@ -3,6 +3,8 @@
 //! Attributes (Section 38.7.1)
 //! The attributes of an enclave are specified by the struct below as described.
 
+use testing::testaso;
+
 bitflags::bitflags! {
     /// Section 38.7.1.
     pub struct Flags: u64 {
