+++ sgx-types/src/lib.rs
@@ -19,60 +19,6 @@ macro_rules! defflags {
     };
 }
 
-macro_rules! testaso {
-    (@off $name:ty=>$field:ident) => {
-        &unsafe { &*core::ptr::null::<$name>() }.$field as *const _ as usize
-    };
-
-    ($(struct $name:ty: $align:expr, $size:expr => { $($field:ident: $offset:expr),* })+) => {
-        #[cfg(test)]
-        #[test]
-        fn align() {
-            use core::mem::align_of;
-
-            $(
-                assert_eq!(
-                    align_of::<$name>(),
-                    $align,
-                    "align: {}",
-                    stringify!($name)
-                );
-            )+
-        }
-
-        #[cfg(test)]
-        #[test]
-        fn size() {
-            use core::mem::size_of;
-
-            $(
-                assert_eq!(
-                    size_of::<$name>(),
-                    $size,
-                    "size: {}",
-                    stringify!($name)
-                );
-            )+
-        }
-
-        #[cfg(test)]
-        #[test]
-        fn offsets() {
-            $(
-                $(
-                    assert_eq!(
-                        testaso!(@off $name=>$field),
-                        $offset,
-                        "offset: {}::{}",
-                        stringify!($name),
-                        stringify!($field)
-                    );
-                )*
-            )+
-        }
-    };
-}
-
 pub mod attr;
 pub mod isv;
 pub mod misc;
