+++ testing/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "sgx-types"
+name = "testing"
 version = "0.1.0"
-authors = ["Nathaniel McCallum <npmccallum@redhat.com>"]
+authors = ["Lily Sturmann <lsturman@redhat.com>", "Nathaniel McCallum <npmccallum@redhat.com>"]
 license = "Apache-2.0"
 edition = "2018"
 
@@ -9,4 +9,3 @@ edition = "2018"
 
 [dependencies]
 bitflags = "1.2.1"
-testing  = { path = "../testing" }
