+++ testing/src/lib.rs
@@ -1,0 +1,71 @@
+// SPDX-License-Identifier: Apache-2.0
+
+//! This crate enables testing for other crates related to Enarx.
+
+#![no_std]
+#![deny(clippy::all)]
+#![allow(clippy::identity_op)]
+#![deny(missing_docs)]
+
+/// This macro enables writing tests for alignment, size, and offset of fields in structs.
+/// Example usage:
+/// testaso! {
+///     struct mystruct: 8, 4096 => {
+///         f1: 0,
+///         f2: 8
+///     }
+/// }
+#[macro_export]
+macro_rules! testaso {
+    (@off $name:ty=>$field:ident) => {
+        &unsafe { &*core::ptr::null::<$name>() }.$field as *const _ as usize
+    };
+
+    ($(struct $name:ty: $align:expr, $size:expr => { $($field:ident: $offset:expr),* })+) => {
+        #[cfg(test)]
+        #[test]
+        fn align() {
+            use core::mem::align_of;
+
+            $(
+                assert_eq!(
+                    align_of::<$name>(),
+                    $align,
+                    "align: {}",
+                    stringify!($name)
+                );
+            )+
+        }
+
+        #[cfg(test)]
+        #[test]
+        fn size() {
+            use core::mem::size_of;
+
+            $(
+                assert_eq!(
+                    size_of::<$name>(),
+                    $size,
+                    "size: {}",
+                    stringify!($name)
+                );
+            )+
+        }
+
+        #[cfg(test)]
+        #[test]
+        fn offsets() {
+            $(
+                $(
+                    assert_eq!(
+                        testaso!(@off $name=>$field),
+                        $offset,
+                        "offset: {}::{}",
+                        stringify!($name),
+                        stringify!($field)
+                    );
+                )*
+ 	    )+
+        }
+    };
+}
